# Update Next.js to version 9.4.4