--- a/packages/web/next.config.js
+++ b/packages/web/next.config.js
@@ -1,24 +1,14 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 const path = require('path');
-const { ANALYZE } = process.env;
+const withBundleAnalyzer = require('@next/bundle-analyzer')({
+	enabled: process.env.ANALYZE === 'true',
+});
 
 // Use next-compose-plugins if this gets weird
-module.exports = {
-	webpack: (config, { isServer }) => {
-		if (ANALYZE) {
-			const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
-
-			config.plugins.push(
-				new BundleAnalyzerPlugin({
-					analyzerMode: 'server',
-					analyzerPort: isServer ? 8888 : 8889,
-					openAnalyzer: true,
-				}),
-			);
-		}
-
+module.exports = withBundleAnalyzer({
+	webpack: config => {
 		config.resolve.alias['@olimat/web'] = path.join(__dirname, 'src');
 
 		return config;
 	},
-};
+});