--- a/packages/web/package.json
+++ b/packages/web/package.json
@@ -32,7 +32,7 @@
     "isomorphic-unfetch": "3.0.0",
     "jss": "10.3.0",
     "lodash": "4.17.15",
-    "next": "^9.1.4",
+    "next": "^9.4.4",
     "notistack": "0.9.17",
     "nprogress": "0.2.0",
     "prop-types": "15.7.2",