Closes #304  #322 

There's a lot of new features:

# [Next.js 9.1.7](https://nextjs.org/blog/next-9-1-7)
Monday, January 6th 2020 (6 months ago)

Next.js 9 was released six (6) months ago, followed by Next.js 9.1 three (3) months ago. These two releases added very powerful new features to Next.js, without increasing our baseline client runtime size.

Since then, we've focused heavily on refining and improving the framework as a whole: 9.1.1, 9.1.2, 9.1.3, 9.1.4, 9.1.5, 9.1.6, and 9.1.7. Let's dive into what these releases have improved!

- 3% – 8%+ Smaller Client-Side JavaScript Size
- Redesigned Production Build CLI Output
- New Built-In Polyfills: fetch(), URL, and Object.assign
- Optimized Page Loading: Better FCP and TTI
- Support for the Latest JavaScript Features
- Zero-Config Deployment Support for next export Applications
- React Strict Mode Compliance and Opt-In
- Automated Testing against Nightly React Builds

# [Next.js 9.2](https://nextjs.org/blog/next-9-2)
Wednesday, January 15th 2020 (5 months ago)

We are excited today to introduce the production-ready Next.js 9.2, featuring:

- Built-In CSS Support for Global Stylesheets
- Built-In CSS Module Support for Component-Level Styles
- **Improved Code-Splitting Strategy**
- Catch-All Dynamic Routes

Highlight:

> **Improved Code-Splitting Strategy**: The Google Chrome team heavily optimized Next.js' code-splitting strategy, resulting in significantly smaller client-side bundles. Furthermore, they've maximized HTTP/2 utilization to improve page load speed without hurting HTTP/1.1 performance.

# [Next.js 9.3](https://nextjs.org/blog/next-9-3)
Monday, March 9th 2020 (3 months ago)

We are excited today to introduce Next.js 9.3, featuring:

- **Next-gen Static Site Generation (SSG) Support**
- Preview Mode
- Built-In Sass Support for Global Stylesheets
- Built-In Sass CSS Module Support for Component-Level Styles
- Automatic Static Optimization for 404
- 32 kB Smaller Runtime
- Next.js Community on GitHub Discussions

Highlight:

> **Next-gen Static Site Generation (SSG) Support**: Built-in optimized static generation through new data fetching methods.

> Today we're incredibly excited to announce two new data fetching methods: `getStaticProps` and `getServerSideProps`. We also include a way to provide parameters to statically generate static pages for dynamic routes: `getStaticPaths`.

# [Next.js 9.4](https://nextjs.org/blog/next-9-4)
Monday, May 11th 2020 (about 1 month ago)

We are excited today to introduce Next.js 9.4, featuring:

- Fast Refresh
- **Incremental Static Regeneration (beta)**
- CMS Examples
- New Environment Variables Support
- Improved Built-in Fetch Support
- Integrated Web Vitals Reporting
- Absolute Imports and Aliases
- Configurable Sass Support
- Improved Log Output

Highlight:

> Next.js introduced Static Site Generation methods in 9.3 with a clear goal in mind: we should get the benefits of static (always fast, always online, globally distributed), but with excellent support for dynamic data, which Next.js is known for.

> To get the best of both worlds, Next.js supports Incremental Static Generation, updating static content after you have already built your site. For example, in 9.3 we’ve introduced the `fallback: true` option in `getStaticPaths`, which allows you to add new pages at runtime.

> Both incremental features (adding pages and lazily updating them), as well as preview mode, are already fully supported by both `next start` and the [Vercel edge platform](https://vercel.com/) out of the box.