@iquabius 2020-06-19
# Improved Code-Splitting Strategy

## Before

```bash
olimat/packages/web on  master [$!?] is 📦 v0.0.1 took 19s 611ms 
➜ npm run analyze

> @olimat/web@0.0.1 analyze /home/iqb/Code/olimat/packages/web
> cross-env ANALYZE=true next build

> Using external babel configuration
> Location: "/home/iqb/Code/olimat/packages/web/.babelrc"
Webpack Bundle Analyzer saved report to /home/iqb/Code/olimat/packages/web/.next/analyze/server.html
Webpack Bundle Analyzer saved report to /home/iqb/Code/olimat/packages/web/.next/analyze/client.html                                    
Creating an optimized production build

Compiled successfully.

Warning: You have opted-out of Automatic Static Optimization due to `getInitialProps` in `pages/_app`.
Read more: https://err.sh/next.js/opt-out-auto-static-optimization

Automatically optimizing pages

Page                        Size
┌ σ /                       11.6 kB
├   /_app                   751 kB
├   /_document
├   /_error                 7.3 kB
├ σ /admin                  4.3 kB
├ σ /admin/cidades          112 kB
├ σ /admin/escolas          108 kB
├ σ /admin/olimpiadas       92.1 kB
├ σ /admin/provas           4.32 kB
├ σ /admin/provas/detalhes  12.2 kB
├ σ /admin/questao          50.9 kB
├ σ /admin/questao-criar    281 kB
├ σ /admin/questao-editar   275 kB
├ σ /admin/questoes         27.6 kB
├ σ /criar_conta            47.4 kB
└ σ /login                  53.8 kB

σ  (Server)       page will be server rendered (i.e. getInitialProps)
*  (Static File)  page was prerendered as static HTML
```

![image](https://user-images.githubusercontent.com/3139777/85178823-7c8c0b00-b24d-11ea-9948-5db293e9c6d0.png)

## After

```bash
olimat/packages/web on  next-v9.4 [$!+?] is 📦 v0.0.1 took 57s 76ms                                                                                                                                                                                  [4/27371]
➜ npm run analyze                                                                                                                                                                                                                                              
                                                                                                                                                                                                                                                               
> @olimat/web@0.0.1 analyze /home/iqb/Code/olimat/packages/web                                                                                                                                                                                                 
> cross-env ANALYZE=true next build                                                                                                                                                                                                                            

info  - Loaded env from /home/iqb/Code/olimat/packages/web/.env
> Using external babel configuration
> Location: "/home/iqb/Code/olimat/packages/web/.babelrc"
Webpack Bundle Analyzer saved report to /home/iqb/Code/olimat/packages/web/.next/analyze/server.html
Webpack Bundle Analyzer saved report to /home/iqb/Code/olimat/packages/web/.next/analyze/client.html
Creating an optimized production build

Compiled successfully.

Warning: You have opted-out of Automatic Static Optimization due to `getInitialProps` in `pages/_app`. This does not opt-out pages with `getStaticProps`
Read more: https://err.sh/next.js/opt-out-auto-static-optimization

Automatically optimizing pages

Page                                                           Size     First Load JS
┌ λ /                                                          4.04 kB         205 kB
├   /_app                                                      22.3 kB         163 kB
├ λ /404                                                       2.57 kB         165 kB
├ λ /admin                                                     1.89 kB         206 kB
├ λ /admin/cidades                                             3.47 kB         235 kB
├ λ /admin/escolas                                             9.28 kB         239 kB
├ λ /admin/olimpiadas                                          3.2 kB          234 kB
├ λ /admin/provas                                              1.9 kB          206 kB
├ λ /admin/provas/detalhes                                     4.28 kB         209 kB
├ λ /admin/questao                                             7.97 kB         217 kB
├ λ /admin/questao-criar                                       3.06 kB         286 kB
├ λ /admin/questao-editar                                      1.19 kB         284 kB
├ λ /admin/questoes                                            7.93 kB         214 kB
├ λ /criar_conta                                               1.93 kB         195 kB
└ λ /login                                                     4.02 kB         197 kB
+ First Load JS shared by all                                  163 kB
  ├ static/pages/_app.js                                       22.3 kB
  ├ chunks/3d66764a5b09324bf2956bd19d9e5cfbf381f96a.8c5faa.js  2.5 kB
  ├ chunks/53896a911918d5eaede5e8178566376a8840e026.de0656.js  6.53 kB
  ├ chunks/7e5ac40d1d8c7fe96d55691e3cf5c27e65153880.9a0128.js  6.56 kB
  ├ chunks/c8fb7b7a52fd2a560f978d2658a6c1bfaec9093e.c58cb5.js  27.4 kB
  ├ chunks/commons.d38219.js                                   10.8 kB
  ├ chunks/df2c40bd6281b2f7f403bfb10872a09e93bb598b.c11759.js  33.7 kB
  ├ chunks/framework.66990f.js                                 45.2 kB
  ├ runtime/main.2eb73a.js                                     6.28 kB
  └ runtime/webpack.aee293.js                                  1.21 kB

λ  (Server)  server-side renders at runtime (uses getInitialProps or getServerSideProps)
○  (Static)  automatically rendered as static HTML (uses no initial props)
●  (SSG)     automatically generated as static HTML + JSON (uses getStaticProps)
```

![image](https://user-images.githubusercontent.com/3139777/85178848-8877cd00-b24d-11ea-9711-e46dd6904316.png)

## @next/bundle-analyzer files

[before-and-after-bundle-analyses.zip](https://github.com/iquabius/olimat/files/4806693/before-and-after-bundle-analyses.zip)